reworked cost structure; compiling but not yet tested

+++ samgods/src/main/java/se/vti/samgods/transportation/consolidation/road/PerformanceMeasures.java
@@ -24,11 +24,11 @@ import org.matsim.api.core.v01.network.Node;
 
 public class PerformanceMeasures {
 
-	public int getTotalArrivalDelay_h(Id<Node> nodeId) {
+	public double getTotalArrivalDelay_h(Id<Node> nodeId) {
 		throw new UnsupportedOperationException();
 	}
 
-	public int getTotalDepartureDelay_h(Id<Node> nodeId) {
+	public double getTotalDepartureDelay_h(Id<Node> nodeId) {
 		throw new UnsupportedOperationException();
 	}
